fix(nextjs): remove the need to install @nx/next for production builds
