+++ packages/next/src/utils/config.spec.ts
@@ -1,7 +1,6 @@
-import type { NextConfig } from 'next';
 import 'nx/src/utils/testing/mock-fs';
-import { composePlugins, createWebpackConfig, NextConfigFn } from './config';
 import { TsconfigPathsPlugin } from 'tsconfig-paths-webpack-plugin';
+import { createWebpackConfig } from './config';
 
 jest.mock('@nx/webpack', () => ({}));
 jest.mock('tsconfig-paths-webpack-plugin');
@@ -74,62 +73,6 @@ describe('Next.js webpack config builder', () => {
       // not much value in checking what they are
       // just check they get added
       expect(config.module.rules.length).toBe(2);
-    });
-  });
-
-  describe('composePlugins', () => {
-    it('should combine multiple plugins', async () => {
-      const nextConfig: NextConfig = {
-        env: {
-          original: 'original',
-        },
-      };
-      const a = (config: NextConfig): NextConfig => {
-        config.env['a'] = 'a';
-        return config;
-      };
-      const b = (config: NextConfig): NextConfig => {
-        config.env['b'] = 'b';
-        return config;
-      };
-      const fn = await composePlugins(a, b);
-      const output = await fn(nextConfig)('test', {});
-
-      expect(output).toEqual({
-        env: {
-          original: 'original',
-          a: 'a',
-          b: 'b',
-        },
-      });
-    });
-
-    it('should compose plugins that return an async function', async () => {
-      const nextConfig: NextConfig = {
-        env: {
-          original: 'original',
-        },
-      };
-      const a = (config: NextConfig): NextConfig => {
-        config.env['a'] = 'a';
-        return config;
-      };
-      const b = (config: NextConfig): NextConfigFn => {
-        return (phase: string) => {
-          config.env['b'] = phase;
-          return config;
-        };
-      };
-      const fn = await composePlugins(a, b);
-      const output = await fn(nextConfig)('test', {});
-
-      expect(output).toEqual({
-        env: {
-          original: 'original',
-          a: 'a',
-          b: 'test',
-        },
-      });
     });
   });
 });
